requirements.txt
telethon
python-dotenv
cryptography
cryptg